requirements.txt
# Core Flask Application
Flask==3.0.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pandas==2.1.4
python-dotenv==1.0.0
openai==1.98.0
anthropic==0.34.2
APScheduler==3.10.4
requests==2.31.0

# Google Analytics 4 Integration
google-analytics-data==0.18.6

# HTML/PDF Report Generation
markdown==3.6
Jinja2==3.1.4
WeasyPrint==66.0

# Production WSGI Server
gunicorn==21.2.0

# Auth
Authlib==1.3.1

# Encryption
cryptography==42.0.8

# Leadgen Dependencies
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6
beautifulsoup4==4.12.2

# Note: WeasyPrint requires system dependencies
# Railway will handle these automatically in production
# Local development: brew install pango gdk-pixbuzz cairo glib